feat(TDQ-15530) Add bijective masking of alphabets with FF1 for Character Operations (#531)
* feat(TDQ-15530) Add alphabet enum and generateFromAlphabet

* Add Bijective Masking to Character operation functions

* Add tests + refactoring

* chore-Fix some sonar code smells

* feat(TDQ-15530) Remove unused method and pass mode a String from Javajet

* feat(TDQ-15530) Factorize setters for FF1  and make Alphabet an attribute of CharacterOperation

* feat(TDQ-15530) Add Function Types + fix codacy issues

* feat(TDQ-15530) Code Review

* feat(TDQ-15530) fix SemanticMaskerTest

* feat(TDQ-15530) Add validation check for length of strings to mask with FF1

* change level of GenrateFromAlphabet's Logger

* feat(TDQ-15530) Fix Serialization issue

+++ dataquality-sampling/src/main/java/org/talend/dataquality/datamasking/generic/fields/FieldDate.java
@@ -72,7 +72,7 @@ public class FieldDate extends AbstractField {
      */
     private void computeDaysPerYear() {
 
-        Integer count = 0;
+        int count = 0;
         for (int year = firstYear; year <= lastYear; year++) {
             numberDaysPerYear.add(count);
             if (isLeapYear(year))
@@ -147,7 +147,7 @@ public class FieldDate extends AbstractField {
             return "";
 
         int year = findNearest(number, numberDaysPerYear);
-        long remainingDays = Integer.valueOf(number - numberDaysPerYear.get(year)).longValue();
+        long remainingDays = number - numberDaysPerYear.get(year);
         int month;
         int days;
         if (isLeapYear(year + firstYear)) {
